move name and description of widget to summaryStep
