+++ MFE/dashboard/src/view/translations/en.dashboard.i18n.json
@@ -75,13 +75,11 @@
     "initial date": "Initial Date",
     "final date": "Final Date"
   },
-  "general": {
+  "summary": {
+    "attributes": "Attributes",
     "name": "Name",
     "description": "Description",
     "nameRequired": "Name is required",
     "enterName": "Enter the name of the panel"
-  },
-  "summary": {
-    "attributes": "Attributes"
   }
 }
