+++ MFE/dashboard/src/view/widget/areaChart/wizard/index.js
@@ -11,17 +11,15 @@ import useArea from '../../wizard/hooks/useArea';
 import {
   Attributes,
   Devices,
-  General,
   Summary,
   Filters,
-  generalValidates,
   attrValidates,
   deviceValidates,
+  summaryValidates,
 } from '../../wizard/Steps';
 import Wizard from '../../wizard/wizard';
 
 const stepsList = [
-  { label: 'steps.general', key: uuidv4() },
   { label: 'steps.devices', key: uuidv4() },
   { label: 'steps.attributes', key: uuidv4() },
   { label: 'steps.filters', key: uuidv4() },
@@ -81,7 +79,6 @@ const WizardPage = ({
       steps={stepsList}
       headerTitle={title}
     >
-      <General validate={generalValidates} name='general' />
       <Devices validate={deviceValidates} name='devices' />
       <Attributes
         validate={attrValidates}
@@ -93,7 +90,7 @@ const WizardPage = ({
         ]}
       />
       <Filters validate={null} name='filters' />
-      <Summary />
+      <Summary validate={summaryValidates} name='summary' />
     </Wizard>
   );
 };
