+++ MFE/dashboard/src/view/widget/table/wizard/index.js
@@ -10,11 +10,10 @@ import { v4 as uuidv4 } from 'uuid';
 import useTable from '../../wizard/hooks/useTable';
 import {
   Attributes,
-  General,
   Summary,
   RealtimeFilter,
-  generalValidates,
   attrValidates,
+  summaryValidates,
 } from '../../wizard/Steps';
 import Selector, {
   selectorValidates,
@@ -22,7 +21,6 @@ import Selector, {
 import Wizard from '../../wizard/wizard';
 
 const stepsList = [
-  { label: 'steps.general', key: uuidv4() },
   { label: 'steps.devices', key: uuidv4() },
   { label: 'steps.attributes', key: uuidv4() },
   { label: 'steps.filters', key: uuidv4() },
@@ -80,7 +78,6 @@ const TableWizard = ({
       headerTitle={title}
       menuState={isMenuOpen}
     >
-      <General validate={generalValidates} name='general' />
       <Selector validate={selectorValidates} />
       <Attributes
         validate={attrValidates}
@@ -89,7 +86,7 @@ const TableWizard = ({
         acceptedTypes={Object.values(TEMPLATE_ATTR_VALUE_TYPES).map(({ value }) => value)}
       />
       <RealtimeFilter validate={null} name='filters' />
-      <Summary />
+      <Summary validate={summaryValidates} name='summary' />
     </Wizard>
   );
 };
