+++ MFE/dashboard/src/view/widget/wizard/Steps/Summary/index.jsx
@@ -1,23 +1,26 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 
 import { Divider, Grid } from '@material-ui/core';
-import { useTranslation } from 'react-i18next';
+import { TextField, makeValidate } from 'mui-rff';
+import { useTranslation, Translation } from 'react-i18next';
+import * as Yup from 'yup';
 import { object2Array } from 'sharedComponents/Utils';
 
 import Wizard from '../../wizard';
 import { useStyles } from './style';
 
-const TitleBox = ({ name, desc }) => {
-  const classes = useStyles();
-  return (
-    <div className={classes.containerText}>
-      <div className={classes.title}>{name}</div>
-      <div className={classes.description}>{desc}</div>
-      <Divider />
-    </div>
-  );
-};
+const schema = Yup.object().shape({
+  name: Yup.string().required('common:required').min(5, 'common:min5characters'),
+});
 
+export const summaryValidates = makeValidate(schema, error => {
+  const { message } = error;
+  return (
+    <Translation key={`t_${message}`}>
+      {t => <span className='error'>{t(`${message}`)}</span>}
+    </Translation>
+  );
+});
 const AttributeBox = ({ name, values = [] }) => {
   const classes = useStyles();
   return (
@@ -39,20 +42,34 @@ const AttributeBox = ({ name, values = [] }) => {
   );
 };
 
-const Summary = ({ values }) => {
+const Summary = ({ validate, name, values }) => {
   const classes = useStyles();
   const { t } = useTranslation(['dashboard']);
 
-  const {
-    general: { name, description },
-    attributes,
-  } = values;
+  const { attributes } = values;
   const attributesList = object2Array(attributes);
   return (
-    <Wizard.Page>
+    <Wizard.Page validate={validate}>
       <Grid container direction='column' className={classes.root}>
+        <Grid className={classes.item} item>
+          <TextField
+            label={t('summary.name')}
+            name='name'
+            variant='outlined'
+            margin='none'
+            fullWidth
+          />
+        </Grid>
+        <Grid className={classes.item} item>
+          <TextField
+            label={t('summary.description')}
+            name='description'
+            variant='outlined'
+            margin='none'
+            fullWidth
+          />
+        </Grid>
         <Grid item className={classes.item}>
-          <TitleBox desc={description} name={name} />
           <AttributeBox name={t('summary.attributes')} values={attributesList} />
         </Grid>
       </Grid>
