+++ MFE/dashboard/src/view/widget/wizard/Steps/index.js
@@ -1,7 +1,6 @@
-export { default as General, generalValidates } from './General';
 export { default as Attributes, attrValidates } from './Attributes';
 export { default as Filters } from './Filters/Filter';
 export { default as RealtimeFilter } from './Filters/RealtimeFilter';
 export { default as Devices, deviceValidates } from './Devices';
 export { Templates } from './Templates';
-export { default as Summary } from './Summary';
+export { default as Summary, summaryValidates } from './Summary';
