+++ MFE/dashboard/src/view/widget/wizard/hooks/useLine.js
@@ -8,11 +8,11 @@ export default (addWidget, addWidgetConfig, addWidgetSaga, generateScheme, addWi
   const lineID = WIDGET.LINE;
 
   const generateLineConfig = useCallback(state => {
-    const { attributes, general: generalState } = state;
+    const { attributes, name, description } = state;
 
     const meta = {
-      title: generalState.name || '',
-      subTitle: generalState.description || '',
+      title: name || '',
+      subTitle: description || '',
     };
 
     const line = object2Array(attributes).map(item => ({
